chore(runtime): keep cleaning the runtime
