+++ app/daemon/src/lib.rs
@@ -5,6 +5,8 @@ pub mod channel;
 pub mod config;
 pub mod feature;
 pub mod gateway;
+pub mod loader;
+pub mod team;
 pub mod utils;
 
 pub use channel::router::{ChannelRouter, RoutingRule};
